12 tools 테스트 추가

+++ BackEnd/tools/views.py
@@ -7,7 +7,7 @@ from django.shortcuts import get_object_or_404
 
 from accounts.decorators import assert_school_code
 from tools.actions import increaseLeft, getLeft, decreaseLeft
-from tools.models import Period, ToolBooking, Tool
+from tools.models import ToolBooking, Tool
 from tools.serializers import ToolSerializer, ToolBookingSerializer
 
 
@@ -20,15 +20,11 @@ class ToolListCreate(ListCreateAPIView):
         return Tool.objects.filter(school=self.request.user.code).order_by('name')
 
     def create(self, request, *args, **kwargs):
-        serializer = self.get_serializer(data=self.request.data)
+        requestData = self.request.data.copy()
+        requestData['school'] = request.user.code
+        serializer = self.get_serializer(data=requestData)
         serializer.is_valid(raise_exception=True)
         data = serializer.validated_data
-
-        if Tool.objects.filter(school=request.user, name=data['name']).exists():
-            return Response(
-                data={'detail': '이미 등록된 이름입니다.'},
-                status=status.HTTP_400_BAD_REQUEST
-            )
 
         Tool.objects.create(
             school=request.user,
@@ -58,52 +54,47 @@ class ToolRetrieveUpdateDestroy(RetrieveUpdateDestroyAPIView):
 @method_decorator(assert_school_code, name='create')
 class ToolBookingCreate(CreateAPIView):
     def create(self, request, *args, **kwargs):
+        tool = get_object_or_404(Tool, **{'school': request.user, 'name': self.request.data['tool']})
         serializer = ToolBookingSerializer(data=self.request.data)
         serializer.is_valid(raise_exception=True)
         data = serializer.validated_data
 
-        tool = get_object_or_404(Tool, **{'school': request.user, 'name': data['tool']})
-
         for i in data['period']:
-            period, created = Period.objects.get_or_create(
-                school=request.user,
-                date=data['date'],
-                period=i,
-                id=data['date'] + '-' + str(i)
-            )
-
-            decreaseLeft(tool, period, data['quantity'])
-
             ToolBooking.objects.create(
                 tool=tool,
-                period=period,
+                date=data['date'],
+                period=int(i),
                 booker=data['booker'],
-                quantity=data['quantity']
+                quantity=data['quantity'],
+                password=data['password']
             )
+            decreaseLeft(tool, data['date'] + '-' + str(i), data['quantity'])
 
-        data = {
-            'school': request.user.name,
+        responseData = {
             'tool': tool.name,
+            'date': data['date'],
             'period': data['period'],
-            'quantity': data['quantity']
+            'booker': data['booker'],
+            'quantity': data['quantity'],
+            'password': data['password']
         }
 
-        return Response(data, status=status.HTTP_201_CREATED)
+        return Response(responseData, status=status.HTTP_201_CREATED)
 
 
 @method_decorator(assert_school_code, name='retrieve')
-class ToolBookingRetrieve(RetrieveAPIView):
+class ToolBookingsByDate(RetrieveAPIView):
     def retrieve(self, request, *args, **kwargs):
         tool = get_object_or_404(Tool, **{'school': request.user, 'name': kwargs['tool']})
-        periods = Period.objects.filter(school=request.user, id__contains=kwargs['date']).order_by('id')
+        bookings = ToolBooking.objects.filter(tool=tool.id, date=kwargs['date'])
 
-        data = {}
-        for period in periods:
-            data[period.period] = list(
-                ToolBooking.objects.filter(
-                    tool=tool.id,
-                    period=period.id
-                ).values('id', 'booker', 'quantity'))
+        data = {1: [], 2: [], 3: [], 4: [], 5: [], 6: []}
+        for booking in bookings:
+            data[booking.period].append({
+                'id': booking.id,
+                'booker': booking.booker,
+                'quantity': booking.quantity
+            })
 
         return Response(data)
 
@@ -112,19 +103,22 @@ class ToolBookingRetrieve(RetrieveAPIView):
 class ToolBookingDestroy(DestroyAPIView):
     def destroy(self, request, *args, **kwargs):
         booking = ToolBooking.objects.get(id=kwargs['bookingId'])
-        increaseLeft(booking.tool, booking.period, booking.quantity)
+
+        if booking.password != request.data['password']:
+            return Response(status=status.HTTP_400_BAD_REQUEST,
+                            data={'detail': '잘못된 비밀번호입니다.'})
+
+        increaseLeft(booking.tool, str(booking.date) + '-' + str(booking.period), booking.quantity)
         booking.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
 
 @method_decorator(assert_school_code, name='retrieve')
-class AvailableLeftRetrieve(RetrieveAPIView):
+class LeftQuantityRetrieve(RetrieveAPIView):
     def retrieve(self, request, *args, **kwargs):
         tool = get_object_or_404(Tool, **{'school': request.user, 'name': kwargs['tool']})
-        periods = Period.objects.filter(school=request.user, id__contains=kwargs['date']).order_by('id')
-
-        lefts = [tool.quantity] * 6
-        for period in periods:
-            lefts[period.period - 1] = getLeft(tool, period)
+        lefts = [0] * 6
+        for i in range(6):
+            lefts[i] = getLeft(tool, kwargs['date'] + '-' + str(i + 1))
 
         return Response(lefts)
